Split metrics to different files.
According to the type of objects for which the metric is collected. The definition of metrics has been moved to the same files.
The backrest_parser.go file contains only general functions.
