+++ backrest/backrest_exporter_metrics_test.go
@@ -1,0 +1,94 @@
+package backrest
+
+import (
+	"bytes"
+	"fmt"
+	"strings"
+	"testing"
+
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/common/expfmt"
+)
+
+func TestGetExporterMetrics(t *testing.T) {
+	type args struct {
+		exporterVer         string
+		testText            string
+		setUpMetricValueFun setUpMetricValueFunType
+	}
+	templateMetrics := `# HELP pgbackrest_exporter_info Information about pgBackRest exporter.
+# TYPE pgbackrest_exporter_info gauge
+pgbackrest_exporter_info{version="unknown"} 1
+`
+	tests := []struct {
+		name string
+		args args
+	}{
+		{"GetExporterInfoGood",
+			args{
+				`unknown`,
+				templateMetrics,
+				setUpMetricValue,
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			getExporterMetrics(tt.args.exporterVer, tt.args.setUpMetricValueFun, logger)
+			reg := prometheus.NewRegistry()
+			reg.MustRegister(pgbrExporterInfoMetric)
+			metricFamily, err := reg.Gather()
+			if err != nil {
+				fmt.Println(err)
+			}
+			out := &bytes.Buffer{}
+			for _, mf := range metricFamily {
+				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
+					panic(err)
+				}
+			}
+			if tt.args.testText != out.String() {
+				t.Errorf("\nVariables do not match:\n%s\nwant:\n%s", tt.args.testText, out.String())
+			}
+		})
+	}
+}
+
+func TestGetExporterInfoErrorsAndDebugs(t *testing.T) {
+	type args struct {
+		exporterVer         string
+		setUpMetricValueFun setUpMetricValueFunType
+		errorsCount         int
+		debugsCount         int
+	}
+	tests := []struct {
+		name string
+		args args
+	}{
+		{"GetExporterInfoLogError",
+			args{
+				`unknown`,
+				fakeSetUpMetricValue,
+				1,
+				1,
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			out := &bytes.Buffer{}
+			logger := log.NewLogfmtLogger(out)
+			lc := log.With(logger, level.AllowInfo())
+			getExporterMetrics(tt.args.exporterVer, tt.args.setUpMetricValueFun, lc)
+			errorsOutputCount := strings.Count(out.String(), "level=error")
+			debugsOutputCount := strings.Count(out.String(), "level=debug")
+			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
+				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
+					tt.args.errorsCount, tt.args.debugsCount,
+					errorsOutputCount, debugsOutputCount)
+			}
+		})
+	}
+}
