+++ backrest/backrest_parser_test.go
@@ -3,7 +3,6 @@ package backrest
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"os"
 	"os/exec"
 	"reflect"
@@ -14,1576 +13,8 @@ import (
 	"time"
 
 	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/expfmt"
 )
-
-type mockBackupLastStruct struct {
-	mockFull mockStruct
-	mockDiff mockStruct
-	mockIncr mockStruct
-}
-
-var (
-	currentUnixTimeForTests = parseDate("2021-07-22 21:00:00").UnixNano()
-	mockDataBackupLast      = mockBackupLastStruct{}
-)
-
-// All metrics exist and all labels are corrected.
-// pgBackrest version = latest.
-func TestGetStanzaMetrics(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		stanzaStatusCode    int
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-	}
-	templateMetrics := `# HELP pgbackrest_stanza_status Current stanza status.
-# TYPE pgbackrest_stanza_status gauge
-pgbackrest_stanza_status{stanza="demo"} 0
-`
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getStanzaMetrics",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Status.Code,
-				setUpMetricValue,
-				templateMetrics,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			getStanzaMetrics(tt.args.stanzaName, tt.args.stanzaStatusCode, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(pgbrStanzaStatusMetric)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf("\nVariables do not match, metrics:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-func TestGetStanzaMetricsErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		stanzaStatusCode    int
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getStanzaMetricsLogError",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Status.Code,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			out := &bytes.Buffer{}
-			lc := log.NewLogfmtLogger(out)
-			getStanzaMetrics(tt.args.stanzaName, tt.args.stanzaStatusCode, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
-}
-
-// All metrics exist and all labels are corrected.
-// pgBackrest version = latest.
-func TestRepoMetrics(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		repoData            []repo
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-	}
-	templateMetrics := `# HELP pgbackrest_repo_status Current repository status.
-# TYPE pgbackrest_repo_status gauge
-pgbackrest_repo_status{cipher="none",repo_key="1",stanza="demo"} 0
-`
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getRepoMetrics",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Repo,
-				setUpMetricValue,
-				templateMetrics,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			getRepoMetrics(tt.args.stanzaName, tt.args.repoData, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(pgbrRepoStatusMetric)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf("\nVariables do not match, metrics:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-func TestGetRepoMetricsErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		repoData            []repo
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getRepoMetricsLogError",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Repo,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			out := &bytes.Buffer{}
-			lc := log.NewLogfmtLogger(out)
-			getRepoMetrics(tt.args.stanzaName, tt.args.repoData, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
-}
-
-// All metrics exist and all labels are corrected.
-// pgBackrest version = latest.
-// With '--backrest.database-count' flag.
-// The case when the backup is performed with block incremental feature flags.
-// Absent metrics:
-//   - pgbackrest_backup_repo_size_bytes
-//
-//nolint:dupl
-func TestGetBackupMetrics(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-		testLastBackups     lastBackupsStruct
-	}
-	templateMetrics := `# HELP pgbackrest_backup_databases Number of databases in backup.
-# TYPE pgbackrest_backup_databases gauge
-pgbackrest_backup_databases{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 1
-# HELP pgbackrest_backup_delta_bytes Amount of data in the database to actually backup.
-# TYPE pgbackrest_backup_delta_bytes gauge
-pgbackrest_backup_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-# HELP pgbackrest_backup_duration_seconds Backup duration.
-# TYPE pgbackrest_backup_duration_seconds gauge
-pgbackrest_backup_duration_seconds{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo",start_time="2021-06-07 09:24:23",stop_time="2021-06-07 09:24:26"} 3
-# HELP pgbackrest_backup_error_status Backup error status.
-# TYPE pgbackrest_backup_error_status gauge
-pgbackrest_backup_error_status{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 1
-# HELP pgbackrest_backup_info Backup info.
-# TYPE pgbackrest_backup_info gauge
-pgbackrest_backup_info{backrest_ver="2.45",backup_name="20210607-092423F",backup_type="full",block_incr="y",database_id="1",lsn_start="0/2000028",lsn_stop="0/2000100",pg_version="13",prior="",repo_key="1",stanza="demo",wal_start="000000010000000000000002",wal_stop="000000010000000000000002"} 1
-# HELP pgbackrest_backup_repo_delta_bytes Compressed files size in backup.
-# TYPE pgbackrest_backup_repo_delta_bytes gauge
-pgbackrest_backup_repo_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_repo_delta_map_bytes Size of block incremental delta map.
-# TYPE pgbackrest_backup_repo_delta_map_bytes gauge
-pgbackrest_backup_repo_delta_map_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 12
-# HELP pgbackrest_backup_repo_size_map_bytes Size of block incremental map.
-# TYPE pgbackrest_backup_repo_size_map_bytes gauge
-pgbackrest_backup_repo_size_map_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 100
-# HELP pgbackrest_backup_size_bytes Full uncompressed size of the database.
-# TYPE pgbackrest_backup_size_bytes gauge
-pgbackrest_backup_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-`
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		{
-			"getBackupMetrics",
-			args{
-				"",
-				"",
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Backup,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				true,
-				setUpMetricValue,
-				templateMetrics,
-				templateLastBackup(),
-			},
-			mockStruct{
-				`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-					`"max":"000000010000000000000002","min":"000000010000000000000001"}],` +
-					`"backup":[{"archive":{"start":"000000010000000000000002","stop":"000000010000000000000002"},` +
-					`"backrest":{"format":5,"version":"2.45"},"database":{"id":1,"repo-key":1},` +
-					`"database-ref":[{"name":"postgres","oid":13412}],"error":true,"error-list":["base/1/3351"],` +
-					`"info":{"delta":24316343,"repository":{"delta":2969512, "delta-map":12,"size-map":100},"size":24316343},` +
-					`"label":"20210614-213200F","lsn":{"start":"0/2000028","stop":"0/2000100"},"prior":null,"reference":null,"timestamp":{"start":1623706320,` +
-					`"stop":1623706322},"type":"full"}],"cipher":"none","db":[{"id":1,"repo-key":1,` +
-					`"system-id":6970977677138971135,"version":"13"}],"name":"demo","repo":[{"cipher":"none",` +
-					`"key":1,"status":{"code":0,"message":"ok"}}],"status":{"code":0,"lock":{"backup":` +
-					`{"held":false}},"message":"ok"}}]`,
-				"",
-				0,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			testLastBackups := getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupInfoMetric,
-				pgbrStanzaBackupDurationMetric,
-				pgbrStanzaBackupDatabaseSizeMetric,
-				pgbrStanzaBackupDatabaseBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMapMetric,
-				pgbrStanzaBackupRepoBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSizeMapMetric,
-				pgbrStanzaBackupErrorMetric,
-				pgbrStanzaBackupDatabasesMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() && !reflect.DeepEqual(tt.args.testLastBackups, testLastBackups) {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s\nlastBackups:\n%v\nwant:\n%v",
-					tt.args.testText, out.String(),
-					tt.args.testLastBackups, testLastBackups,
-				)
-			}
-		})
-	}
-}
-
-// Absent metrics:
-// 	- pgbackrest_backup_repo_size_map_bytes
-//  - pgbackrest_backup_repo_delta_map_bytes
-//
-// pgBackrest version < 2.44.
-//
-//nolint:dupl
-
-func TestGetRepoMapMetricsAbsent(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-		testLastBackups     lastBackupsStruct
-	}
-	templateMetrics := `# HELP pgbackrest_backup_delta_bytes Amount of data in the database to actually backup.
-# TYPE pgbackrest_backup_delta_bytes gauge
-pgbackrest_backup_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-# HELP pgbackrest_backup_duration_seconds Backup duration.
-# TYPE pgbackrest_backup_duration_seconds gauge
-pgbackrest_backup_duration_seconds{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo",start_time="2021-06-07 09:24:23",stop_time="2021-06-07 09:24:26"} 3
-# HELP pgbackrest_backup_error_status Backup error status.
-# TYPE pgbackrest_backup_error_status gauge
-pgbackrest_backup_error_status{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 1
-# HELP pgbackrest_backup_info Backup info.
-# TYPE pgbackrest_backup_info gauge
-pgbackrest_backup_info{backrest_ver="2.41",backup_name="20210607-092423F",backup_type="full",block_incr="n",database_id="1",lsn_start="0/2000028",lsn_stop="0/2000100",pg_version="13",prior="",repo_key="1",stanza="demo",wal_start="000000010000000000000002",wal_stop="000000010000000000000002"} 1
-# HELP pgbackrest_backup_repo_delta_bytes Compressed files size in backup.
-# TYPE pgbackrest_backup_repo_delta_bytes gauge
-pgbackrest_backup_repo_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_repo_size_bytes Full compressed files size to restore the database from backup.
-# TYPE pgbackrest_backup_repo_size_bytes gauge
-pgbackrest_backup_repo_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_size_bytes Full uncompressed size of the database.
-# TYPE pgbackrest_backup_size_bytes gauge
-pgbackrest_backup_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-`
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		{
-			"getBackupMetrics",
-			args{
-				"",
-				"",
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true, 2969514).Name,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true, 2969514).Backup,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true, 2969514).DB,
-				true,
-				setUpMetricValue,
-				templateMetrics,
-				templateLastBackup(),
-			},
-			mockStruct{
-				"",
-				"ERROR: [031]: invalid option '--repo1-block'",
-				31,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			testLastBackups := getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupInfoMetric,
-				pgbrStanzaBackupDurationMetric,
-				pgbrStanzaBackupDatabaseSizeMetric,
-				pgbrStanzaBackupDatabaseBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMetric,
-				pgbrStanzaBackupRepoBackupSizeMetric,
-				pgbrStanzaBackupErrorMetric,
-				pgbrStanzaBackupDatabasesMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() && !reflect.DeepEqual(tt.args.testLastBackups, testLastBackups) {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s\nlastBackups:\n%v\nwant:\n%v",
-					tt.args.testText, out.String(),
-					tt.args.testLastBackups, testLastBackups,
-				)
-			}
-		})
-	}
-}
-
-// Absent metrics:
-//   - pgbackrest_backup_databases
-//   - pgbackrest_backup_repo_size_map_bytes
-//   - pgbackrest_backup_repo_delta_map_bytes
-//
-// pgBackrest version < 2.41.
-//
-//nolint:dupl
-func TestGetBackupMetricsDBsAbsent(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-		testLastBackups     lastBackupsStruct
-	}
-	templateMetrics := `# HELP pgbackrest_backup_delta_bytes Amount of data in the database to actually backup.
-# TYPE pgbackrest_backup_delta_bytes gauge
-pgbackrest_backup_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-# HELP pgbackrest_backup_duration_seconds Backup duration.
-# TYPE pgbackrest_backup_duration_seconds gauge
-pgbackrest_backup_duration_seconds{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo",start_time="2021-06-07 09:24:23",stop_time="2021-06-07 09:24:26"} 3
-# HELP pgbackrest_backup_error_status Backup error status.
-# TYPE pgbackrest_backup_error_status gauge
-pgbackrest_backup_error_status{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 1
-# HELP pgbackrest_backup_info Backup info.
-# TYPE pgbackrest_backup_info gauge
-pgbackrest_backup_info{backrest_ver="2.41",backup_name="20210607-092423F",backup_type="full",block_incr="n",database_id="1",lsn_start="0/2000028",lsn_stop="0/2000100",pg_version="13",prior="",repo_key="1",stanza="demo",wal_start="000000010000000000000002",wal_stop="000000010000000000000002"} 1
-# HELP pgbackrest_backup_repo_delta_bytes Compressed files size in backup.
-# TYPE pgbackrest_backup_repo_delta_bytes gauge
-pgbackrest_backup_repo_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_repo_size_bytes Full compressed files size to restore the database from backup.
-# TYPE pgbackrest_backup_repo_size_bytes gauge
-pgbackrest_backup_repo_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_size_bytes Full uncompressed size of the database.
-# TYPE pgbackrest_backup_size_bytes gauge
-pgbackrest_backup_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-`
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		{
-			"getBackupMetrics",
-			args{
-				"",
-				"",
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).Name,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).Backup,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).DB,
-				true,
-				setUpMetricValue,
-				templateMetrics,
-				templateLastBackup(),
-			},
-			mockStruct{
-				"",
-				"ERROR: [027]: option 'set' is currently only valid for text output",
-				27,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			testLastBackups := getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupInfoMetric,
-				pgbrStanzaBackupDurationMetric,
-				pgbrStanzaBackupDatabaseSizeMetric,
-				pgbrStanzaBackupDatabaseBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMetric,
-				pgbrStanzaBackupRepoBackupSizeMetric,
-				pgbrStanzaBackupErrorMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() && !reflect.DeepEqual(tt.args.testLastBackups, testLastBackups) {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s\nlastBackups:\n%v\nwant:\n%v",
-					tt.args.testText, out.String(),
-					tt.args.testLastBackups, testLastBackups,
-				)
-			}
-		})
-	}
-}
-
-// Absent metrics:
-//   - pgbackrest_backup_error_status
-//   - pgbackrest_backup_databases
-//   - pgbackrest_backup_repo_size_map_bytes
-//   - pgbackrest_backup_repo_delta_map_bytes
-//
-// Labels:
-//   - lsn_start=""
-//   - lsn_stop=""
-//
-// pgBackrest version < 2.36.
-//
-//nolint:dupl
-func TestGetBackupMetricsErrorAbsent(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-		testLastBackups     lastBackupsStruct
-	}
-	templateMetrics := `# HELP pgbackrest_backup_delta_bytes Amount of data in the database to actually backup.
-# TYPE pgbackrest_backup_delta_bytes gauge
-pgbackrest_backup_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-# HELP pgbackrest_backup_duration_seconds Backup duration.
-# TYPE pgbackrest_backup_duration_seconds gauge
-pgbackrest_backup_duration_seconds{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo",start_time="2021-06-07 09:24:23",stop_time="2021-06-07 09:24:26"} 3
-# HELP pgbackrest_backup_info Backup info.
-# TYPE pgbackrest_backup_info gauge
-pgbackrest_backup_info{backrest_ver="2.35",backup_name="20210607-092423F",backup_type="full",block_incr="n",database_id="1",lsn_start="",lsn_stop="",pg_version="13",prior="",repo_key="1",stanza="demo",wal_start="000000010000000000000002",wal_stop="000000010000000000000002"} 1
-# HELP pgbackrest_backup_repo_delta_bytes Compressed files size in backup.
-# TYPE pgbackrest_backup_repo_delta_bytes gauge
-pgbackrest_backup_repo_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_repo_size_bytes Full compressed files size to restore the database from backup.
-# TYPE pgbackrest_backup_repo_size_bytes gauge
-pgbackrest_backup_repo_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_size_bytes Full uncompressed size of the database.
-# TYPE pgbackrest_backup_size_bytes gauge
-pgbackrest_backup_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="1",stanza="demo"} 2.4316343e+07
-`
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		{
-			"getBackupMetricsErrorAbsent",
-			args{
-				"",
-				"",
-				templateStanzaErrorAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).Name,
-				templateStanzaErrorAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).Backup,
-				templateStanzaErrorAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).DB,
-				true,
-				setUpMetricValue,
-				templateMetrics,
-				templateLastBackup(),
-			},
-			mockStruct{
-				"",
-				"ERROR: [027]: option 'set' is currently only valid for text output",
-				27,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			testLastBackups := getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupInfoMetric,
-				pgbrStanzaBackupDurationMetric,
-				pgbrStanzaBackupDatabaseSizeMetric,
-				pgbrStanzaBackupDatabaseBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMetric,
-				pgbrStanzaBackupRepoBackupSizeMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() && !reflect.DeepEqual(tt.args.testLastBackups, testLastBackups) {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s\nlastBackups:\n%v\nwant:\n%v",
-					tt.args.testText, out.String(),
-					tt.args.testLastBackups, testLastBackups,
-				)
-			}
-		})
-	}
-}
-
-// Absent metrics:
-//   - pgbackrest_backup_error_status
-//   - pgbackrest_backup_databases
-//   - pgbackrest_backup_repo_size_map_bytes
-//   - pgbackrest_backup_repo_delta_map_bytes
-//
-// Labels:
-//   - repo_key="0"
-//   - lsn_start=""
-//   - lsn_stop=""
-//
-// pgBackrest version < v2.32
-//
-//nolint:dupl
-func TestGetBackupMetricsRepoAbsent(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-		testLastBackups     lastBackupsStruct
-	}
-	templateMetrics := `# HELP pgbackrest_backup_delta_bytes Amount of data in the database to actually backup.
-# TYPE pgbackrest_backup_delta_bytes gauge
-pgbackrest_backup_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="0",stanza="demo"} 2.4316343e+07
-# HELP pgbackrest_backup_duration_seconds Backup duration.
-# TYPE pgbackrest_backup_duration_seconds gauge
-pgbackrest_backup_duration_seconds{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="0",stanza="demo",start_time="2021-06-07 09:24:23",stop_time="2021-06-07 09:24:26"} 3
-# HELP pgbackrest_backup_info Backup info.
-# TYPE pgbackrest_backup_info gauge
-pgbackrest_backup_info{backrest_ver="2.31",backup_name="20210607-092423F",backup_type="full",block_incr="n",database_id="1",lsn_start="",lsn_stop="",pg_version="13",prior="",repo_key="0",stanza="demo",wal_start="000000010000000000000002",wal_stop="000000010000000000000002"} 1
-# HELP pgbackrest_backup_repo_delta_bytes Compressed files size in backup.
-# TYPE pgbackrest_backup_repo_delta_bytes gauge
-pgbackrest_backup_repo_delta_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="0",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_repo_size_bytes Full compressed files size to restore the database from backup.
-# TYPE pgbackrest_backup_repo_size_bytes gauge
-pgbackrest_backup_repo_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="0",stanza="demo"} 2.969514e+06
-# HELP pgbackrest_backup_size_bytes Full uncompressed size of the database.
-# TYPE pgbackrest_backup_size_bytes gauge
-pgbackrest_backup_size_bytes{backup_name="20210607-092423F",backup_type="full",database_id="1",repo_key="0",stanza="demo"} 2.4316343e+07
-`
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		{
-			"getBackupMetricsRepoAbsent",
-			args{
-				"",
-				"",
-				templateStanzaRepoAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).Name,
-				templateStanzaRepoAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).Backup,
-				templateStanzaRepoAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					2969514).DB,
-				false,
-				setUpMetricValue,
-				templateMetrics,
-				templateLastBackup(),
-			},
-			mockStruct{
-				"",
-				"ERROR: [027]: option 'set' is currently only valid for text output",
-				27,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			testLastBackups := getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupInfoMetric,
-				pgbrStanzaBackupDurationMetric,
-				pgbrStanzaBackupDatabaseSizeMetric,
-				pgbrStanzaBackupDatabaseBackupSizeMetric,
-				pgbrStanzaBackupRepoBackupSetSizeMetric,
-				pgbrStanzaBackupRepoBackupSizeMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() && !reflect.DeepEqual(tt.args.testLastBackups, testLastBackups) {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s\nlastBackups:\n%v\nwant:\n%v",
-					tt.args.testText, out.String(),
-					tt.args.testLastBackups, testLastBackups,
-				)
-			}
-		})
-	}
-}
-
-func TestGetBackupMetricsErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		backupData          []backup
-		dbData              []db
-		backupDBCount       bool
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name         string
-		args         args
-		mockTestData mockStruct
-	}{
-		// pgBackRest >= 2.45
-		// Without backup set size.
-		{
-			"getBackupMetricsLogError",
-			args{
-				"",
-				"",
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Backup,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				true,
-				fakeSetUpMetricValue,
-				9,
-				9,
-			},
-			mockStruct{
-				`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-					`"max":"000000010000000000000002","min":"000000010000000000000001"}],` +
-					`"backup":[{"archive":{"start":"000000010000000000000002","stop":"000000010000000000000002"},` +
-					`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-					`"database-ref":[{"name":"postgres","oid":13412}],"error":true,"error-list":["base/1/3351"],` +
-					`"info":{"delta":24316343,"repository":{"delta":2969512, "delta-map":12,"size-map":100},"size":24316343},` +
-					`"label":"20210614-213200F","lsn":{"start":"0/2000028","stop":"0/2000100"},"prior":null,"reference":null,"timestamp":{"start":1623706320,` +
-					`"stop":1623706322},"type":"full"}],"cipher":"none","db":[{"id":1,"repo-key":1,` +
-					`"system-id":6970977677138971135,"version":"13"}],"name":"demo","repo":[{"cipher":"none",` +
-					`"key":1,"status":{"code":0,"message":"ok"}}],"status":{"code":0,"lock":{"backup":` +
-					`{"held":false}},"message":"ok"}}]`,
-				"",
-				0,
-			},
-		},
-		// pgBackrest older than v2.45.
-		// Here the version is not important.
-		// Getting an error is being tested for backup set size.
-		{
-			"getBackupMetricsLogErrorWithRepo",
-			args{
-				"",
-				"",
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).Name,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).Backup,
-				templateStanzaRepoMapSizesAbsent(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					2969514).DB,
-				true,
-				fakeSetUpMetricValue,
-				8,
-				8,
-			},
-			mockStruct{
-				`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-					`"max":"000000010000000000000002","min":"000000010000000000000001"}],` +
-					`"backup":[{"archive":{"start":"000000010000000000000002","stop":"000000010000000000000002"},` +
-					`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-					`"database-ref":[{"name":"postgres","oid":13412}],"error":true,"error-list":["base/1/3351"],` +
-					`"info":{"delta":24316343,"repository":{"delta":2969512,"size":2969512},"size":24316343},` +
-					`"label":"20210614-213200F","lsn":{"start":"0/2000028","stop":"0/2000100"},"prior":null,"reference":null,"timestamp":{"start":1623706320,` +
-					`"stop":1623706322},"type":"full"}],"cipher":"none","db":[{"id":1,"repo-key":1,` +
-					`"system-id":6970977677138971135,"version":"13"}],"name":"demo","repo":[{"cipher":"none",` +
-					`"key":1,"status":{"code":0,"message":"ok"}}],"status":{"code":0,"lock":{"backup":` +
-					`{"held":false}},"message":"ok"}}]`,
-				"",
-				0,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockData = tt.mockTestData
-			execCommand = fakeExecCommand
-			defer func() { execCommand = exec.Command }()
-			out := &bytes.Buffer{}
-			lc := log.NewLogfmtLogger(out)
-			getBackupMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.backupData, tt.args.dbData, tt.args.backupDBCount, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
-}
-
-// All metrics exist and all labels are corrected.
-// pgBackrest version = latest.
-// With '--backrest.database-count-latest' flag.
-func TestGetBackupLastMetrics(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		lastBackups         lastBackupsStruct
-		backupDBCountLatest bool
-		currentUnixTime     int64
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-	}
-	templateMetrics := `# HELP pgbackrest_backup_last_databases Number of databases in the last full, differential or incremental backup.
-# TYPE pgbackrest_backup_last_databases gauge
-pgbackrest_backup_last_databases{backup_type="diff",stanza="demo"} 2
-pgbackrest_backup_last_databases{backup_type="full",stanza="demo"} 1
-pgbackrest_backup_last_databases{backup_type="incr",stanza="demo"} 2
-# HELP pgbackrest_backup_since_last_completion_seconds Seconds since the last completed full, differential or incremental backup.
-# TYPE pgbackrest_backup_since_last_completion_seconds gauge
-pgbackrest_backup_since_last_completion_seconds{backup_type="diff",stanza="demo"} 9.223372036854776e+09
-pgbackrest_backup_since_last_completion_seconds{backup_type="full",stanza="demo"} 9.223372036854776e+09
-pgbackrest_backup_since_last_completion_seconds{backup_type="incr",stanza="demo"} 9.223372036854776e+09
-`
-	tests := []struct {
-		name                   string
-		args                   args
-		mockTestDataBackupLast mockBackupLastStruct
-	}{
-		{
-			"getBackupLastMetrics",
-			args{
-				"",
-				"",
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateLastBackupDifferent(),
-				true,
-				currentUnixTimeForTests,
-				setUpMetricValue,
-				templateMetrics,
-			},
-			mockBackupLastStruct{
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000003","stop":"000000010000000000000004"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-						`"database-ref":[{"name":"postgres","oid":13412}],"error":false,` +
-						`"info":{"delta":32230330,"repository":{"delta":3970793,"size":3970793},"size":32230330},` +
-						`"label":"20220926-201857F","link":null,"lsn":{"start":"0/3000028","stop":"0/4000050"},"prior":null,` +
-						`"reference":null,"tablespace":null,"timestamp":{"start":1664223537,"stop":1664223540},"type":"full"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000005","stop":"000000010000000000000006"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":2},` +
-						`"database-ref":[{"name":"postgres","oid":13412},{"name":"test_db","oid":16384}],` +
-						`"error":false,"info":{"delta":16657,"repository":{"delta":518,"size":3970802},"size":32230338},` +
-						`"label":"20220926-201857F_20220926-201901D","link":null,"lsn":{"start":"0/5000028","stop":"0/6000050"},"prior":"20220926-201857F",` +
-						`"reference":["20220926-201857F"],"tablespace":null,"timestamp":{"start":1664223541,"stop":1664223543},"type":"diff"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000008","stop":"000000010000000000000008"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-						`"database-ref":[{"name":"postgres","oid":13412},{"name":"test_db","oid":16384}],` +
-						`"error":false,"info":{"delta":16657,"repository":{"delta":519,"size":3970803},"size":32230338},` +
-						`"label":"20220926-201854F_20220926-202454I","link":null,"lsn":{"start":"0/8000028","stop":"0/8000138"},` +
-						`"prior":"20220926-201854F","reference":["20220926-201854F"],"tablespace":null,` +
-						`"timestamp":{"start":1664223894,"stop":1664223896},"type":"incr"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockDataBackupLast = tt.mockTestDataBackupLast
-			execCommand = fakeExecCommandSpecificDatabase
-			defer func() { execCommand = exec.Command }()
-			lc := log.NewNopLogger()
-			getBackupLastMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.lastBackups, tt.args.backupDBCountLatest, tt.args.currentUnixTime, tt.args.setUpMetricValueFun, lc)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupSinceLastCompletionSecondsMetric,
-				pgbrStanzaBackupLastDatabasesMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-// Absent metrics:
-//   - pgbackrest_backup_last_databases.
-//
-// pgBackrest version < v2.41.
-func TestGetBackupLastMetricsDBsAbsent(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		lastBackups         lastBackupsStruct
-		backupDBCountLatest bool
-		currentUnixTime     int64
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-	}
-	templateMetrics := `# HELP pgbackrest_backup_since_last_completion_seconds Seconds since the last completed full, differential or incremental backup.
-# TYPE pgbackrest_backup_since_last_completion_seconds gauge
-pgbackrest_backup_since_last_completion_seconds{backup_type="diff",stanza="demo"} 9.223372036854776e+09
-pgbackrest_backup_since_last_completion_seconds{backup_type="full",stanza="demo"} 9.223372036854776e+09
-pgbackrest_backup_since_last_completion_seconds{backup_type="incr",stanza="demo"} 9.223372036854776e+09
-`
-	tests := []struct {
-		name                   string
-		args                   args
-		mockTestDataBackupLast mockBackupLastStruct
-	}{
-		{
-			"getBackupLastMetrics",
-			args{
-				"",
-				"",
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateLastBackupDifferent(),
-				true,
-				currentUnixTimeForTests,
-				setUpMetricValue,
-				templateMetrics,
-			},
-			mockBackupLastStruct{
-				mockStruct{
-					"",
-					"ERROR: [027]: option 'set' is currently only valid for text output",
-					27,
-				},
-				mockStruct{
-					"",
-					"ERROR: [027]: option 'set' is currently only valid for text output",
-					27,
-				},
-				mockStruct{
-					"",
-					"ERROR: [027]: option 'set' is currently only valid for text output",
-					27,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			mockDataBackupLast = tt.mockTestDataBackupLast
-			execCommand = fakeExecCommandSpecificDatabase
-			defer func() { execCommand = exec.Command }()
-			lc := log.NewNopLogger()
-			getBackupLastMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.lastBackups, tt.args.backupDBCountLatest, tt.args.currentUnixTime, tt.args.setUpMetricValueFun, lc)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(
-				pgbrStanzaBackupSinceLastCompletionSecondsMetric,
-			)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-func TestGetBackupLastMetricsErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		config              string
-		configIncludePath   string
-		stanzaName          string
-		lastBackups         lastBackupsStruct
-		backupDBCountLatest bool
-		currentUnixTime     int64
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name                   string
-		args                   args
-		mockTestDataBackupLast mockBackupLastStruct
-	}{
-		{
-			"getBackupLastMetricsLogError",
-			args{
-				"",
-				"",
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateLastBackup(),
-				true,
-				currentUnixTimeForTests,
-				fakeSetUpMetricValue,
-				6,
-				6,
-			},
-			mockBackupLastStruct{
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000003","stop":"000000010000000000000004"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-						`"database-ref":[{"name":"postgres","oid":13412}],"error":false,` +
-						`"info":{"delta":32230330,"repository":{"delta":3970793,"size":3970793},"size":32230330},` +
-						`"label":"20220926-201857F","link":null,"lsn":{"start":"0/3000028","stop":"0/4000050"},"prior":null,` +
-						`"reference":null,"tablespace":null,"timestamp":{"start":1664223537,"stop":1664223540},"type":"full"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000005","stop":"000000010000000000000006"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":2},` +
-						`"database-ref":[{"name":"postgres","oid":13412},{"name":"test_db","oid":16384}],` +
-						`"error":false,"info":{"delta":16657,"repository":{"delta":518,"size":3970802},"size":32230338},` +
-						`"label":"20220926-201857F_20220926-201901D","link":null,"lsn":{"start":"0/5000028","stop":"0/6000050"},"prior":"20220926-201857F",` +
-						`"reference":["20220926-201857F"],"tablespace":null,"timestamp":{"start":1664223541,"stop":1664223543},"type":"diff"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-				mockStruct{
-					`[{"archive":[{"database":{"id":1,"repo-key":1},"id":"13-1",` +
-						`"max":"000000010000000000000010","min":"000000010000000000000001"}],` +
-						`"backup":[{"archive":{"start":"000000010000000000000008","stop":"000000010000000000000008"},` +
-						`"backrest":{"format":5,"version":"2.41"},"database":{"id":1,"repo-key":1},` +
-						`"database-ref":[{"name":"postgres","oid":13412},{"name":"test_db","oid":16384}],` +
-						`"error":false,"info":{"delta":16657,"repository":{"delta":519,"size":3970803},"size":32230338},` +
-						`"label":"20220926-201854F_20220926-202454I","link":null,"lsn":{"start":"0/8000028","stop":"0/8000138"},` +
-						`"prior":"20220926-201854F","reference":["20220926-201854F"],"tablespace":null,` +
-						`"timestamp":{"start":1664223894,"stop":1664223896},"type":"incr"}],` +
-						`"cipher":"none","db":[{"id":1,"repo-key":1,"system-id":7147741414128675215,"version":"13"}],` +
-						`"name":"demo","repo":[{"cipher":"none","key":1,"status":{"code":0,"message":"ok"}}],` +
-						`"status":{"code":0,"lock":{"backup":{"held":false}},"message":"ok"}}]`,
-					"",
-					0,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockDataBackupLast = tt.mockTestDataBackupLast
-			execCommand = fakeExecCommandSpecificDatabase
-			defer func() { execCommand = exec.Command }()
-			out := &bytes.Buffer{}
-			lc := log.NewLogfmtLogger(out)
-			getBackupLastMetrics(tt.args.config, tt.args.configIncludePath, tt.args.stanzaName, tt.args.lastBackups, tt.args.backupDBCountLatest, tt.args.currentUnixTime, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
-}
-
-// All metrics exist and all labels are corrected.
-// pgBackrest version = latest.
-func TestGetWALMetrics(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		archiveData         []archive
-		dbData              []db
-		verboseWAL          bool
-		setUpMetricValueFun setUpMetricValueFunType
-		testText            string
-	}
-	templateMetrics := `# HELP pgbackrest_wal_archive_status Current WAL archive status.
-# TYPE pgbackrest_wal_archive_status gauge
-`
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getWALMetricsVerboseFalse",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				false,
-				setUpMetricValue,
-				templateMetrics +
-					`pgbackrest_wal_archive_status{database_id="1",pg_version="13",repo_key="1",stanza="demo",wal_max="",wal_min=""} 1` +
-					"\n",
-			},
-		},
-		{
-			"getWALMetricsVerboseTrue",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				true,
-				setUpMetricValue,
-				templateMetrics +
-					`pgbackrest_wal_archive_status{database_id="1",pg_version="13",repo_key="1",stanza="demo",wal_max="000000010000000000000004",wal_min="000000010000000000000001"} 1` +
-					"\n",
-			},
-		},
-		{
-			"getWALMetricsWithoutWAL",
-			args{
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				false,
-				setUpMetricValue,
-				templateMetrics +
-					`pgbackrest_wal_archive_status{database_id="1",pg_version="13",repo_key="1",stanza="demo",wal_max="",wal_min=""} 0` +
-					"\n",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetMetrics()
-			getWALMetrics(tt.args.stanzaName, tt.args.archiveData, tt.args.dbData, tt.args.verboseWAL, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(pgbrWALArchivingMetric)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf(
-					"\nVariables do not match, metrics:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-func TestGetWALMetricsErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		stanzaName          string
-		archiveData         []archive
-		dbData              []db
-		verboseWAL          bool
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			"getWALMetricsVerboseFalseLogError",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				false,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-		{
-			"getMetricsVerboseTrueLogError",
-			args{
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"000000010000000000000004",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				true,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-		{
-			"getMetricsWithoutWalLogError",
-			args{
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Name,
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).Archive,
-				templateStanza(
-					"",
-					"000000010000000000000001",
-					[]databaseRef{{"postgres", 13425}},
-					true,
-					12,
-					100).DB,
-				false,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			out := &bytes.Buffer{}
-			lc := log.NewLogfmtLogger(out)
-			getWALMetrics(tt.args.stanzaName, tt.args.archiveData, tt.args.dbData, tt.args.verboseWAL, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
-}
 
 func TestGetPGVersion(t *testing.T) {
 	type args struct {
@@ -2235,87 +666,6 @@ func parseDate(value string) time.Time {
 		panic(err)
 	}
 	return valueReturn
-}
-
-func TestGetExporterMetrics(t *testing.T) {
-	type args struct {
-		exporterVer         string
-		testText            string
-		setUpMetricValueFun setUpMetricValueFunType
-	}
-	templateMetrics := `# HELP pgbackrest_exporter_info Information about pgBackRest exporter.
-# TYPE pgbackrest_exporter_info gauge
-pgbackrest_exporter_info{version="unknown"} 1
-`
-	tests := []struct {
-		name string
-		args args
-	}{
-		{"GetExporterInfoGood",
-			args{
-				`unknown`,
-				templateMetrics,
-				setUpMetricValue,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			getExporterMetrics(tt.args.exporterVer, tt.args.setUpMetricValueFun, logger)
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(pgbrExporterInfoMetric)
-			metricFamily, err := reg.Gather()
-			if err != nil {
-				fmt.Println(err)
-			}
-			out := &bytes.Buffer{}
-			for _, mf := range metricFamily {
-				if _, err := expfmt.MetricFamilyToText(out, mf); err != nil {
-					panic(err)
-				}
-			}
-			if tt.args.testText != out.String() {
-				t.Errorf("\nVariables do not match:\n%s\nwant:\n%s", tt.args.testText, out.String())
-			}
-		})
-	}
-}
-
-func TestGetExporterInfoErrorsAndDebugs(t *testing.T) {
-	type args struct {
-		exporterVer         string
-		setUpMetricValueFun setUpMetricValueFunType
-		errorsCount         int
-		debugsCount         int
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{"GetExporterInfoLogError",
-			args{
-				`unknown`,
-				fakeSetUpMetricValue,
-				1,
-				1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			out := &bytes.Buffer{}
-			logger := log.NewLogfmtLogger(out)
-			lc := log.With(logger, level.AllowInfo())
-			getExporterMetrics(tt.args.exporterVer, tt.args.setUpMetricValueFun, lc)
-			errorsOutputCount := strings.Count(out.String(), "level=error")
-			debugsOutputCount := strings.Count(out.String(), "level=debug")
-			if tt.args.errorsCount != errorsOutputCount || tt.args.debugsCount != debugsOutputCount {
-				t.Errorf("\nVariables do not match:\nerrors=%d, debugs=%d\nwant:\nerrors=%d, debugs=%d",
-					tt.args.errorsCount, tt.args.debugsCount,
-					errorsOutputCount, debugsOutputCount)
-			}
-		})
-	}
 }
 
 func TestConvertBoolToFloat64(t *testing.T) {
